+++ core/CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file. This projec
 
 + [#1814](https://github.com/luyadev/luya/issues/1814) Added new json behavior in order to encode/deocde given class attributes.
 + [#1809](https://github.com/luyadev/luya/issues/1809) Add new `autoFormat` in formatter component in order to format values based on the input type.
++ [#1815](https://github.com/luyadev/luya/pull/1815) Add new `TelephoneLink` class to support html anchor with "tel:".
 
 ### Changed